+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/Checkout/Checkout+ApplePay.swift
@@ -368,4 +368,43 @@ extension STPApplePayContext {
             postalCode: postalAddress.postalCode.nonEmpty
         )
     }
+
+    /// Converts default billing details into a `PKContact` for pre-populating the Apple Pay sheet.
+    static func makeBillingContact(from billingDetails: PaymentSheet.BillingDetails) -> PKContact {
+        let contact = PKContact()
+
+        if let name = billingDetails.name {
+            contact.name = PersonNameComponentsFormatter().personNameComponents(from: name)
+        }
+
+        if let phone = billingDetails.phone {
+            contact.phoneNumber = CNPhoneNumber(stringValue: phone)
+        }
+
+        let postalAddress = CNMutablePostalAddress()
+        let address = billingDetails.address
+        postalAddress.isoCountryCode = address.country ?? ""
+
+        var streetComponents: [String] = []
+        if let line1 = address.line1 { streetComponents.append(line1) }
+        if let line2 = address.line2 { streetComponents.append(line2) }
+        if !streetComponents.isEmpty {
+            postalAddress.street = streetComponents.joined(separator: "\n")
+        }
+
+        if let city = address.city {
+            postalAddress.city = city
+        }
+
+        if let state = address.state {
+            postalAddress.state = state
+        }
+
+        if let postalCode = address.postalCode {
+            postalAddress.postalCode = postalCode
+        }
+
+        contact.postalAddress = postalAddress
+        return contact
+    }
 }
